Move THIR structure definitions to `rustc_middle`
